+++ packages/dumb/src/components/tooltips/events-superposition-popover/events-superposition-popover.tsx
@@ -43,7 +43,7 @@ import { computeDateValue, getDateTitleForBaseDatum } from '../../../utils/toolt
 import { getAlarmEventIcon, getAlarmEventTitle } from '../../../utils/alarm-event/alarm-event.util'
 import { useTranslation } from 'react-i18next'
 import { getReservoirChangeIcon, getReservoirChangeTitle } from '../../../utils/reservoir-change/reservoir-change.util'
-import dexcomG6Icon from 'dexcom-g6.svg'
+import warmupIcon from 'warmup-icon.svg'
 import parameterChangeIcon from 'parameter-change.png'
 import styles from './events-superposition-popover.css'
 import { BgPrefs } from '../../../models/blood-glucose.model'
@@ -119,8 +119,7 @@ const EventsSuperpositionPopover: FC<EventsSuperpositionPopoverProps> = (props) 
         return getReservoirChangeIcon(pumpManufacturer)
       }
       case DeviceEventSubtype.Warmup:
-        // TODO to update
-        return dexcomG6Icon
+        return warmupIcon
       default:
         return ''
     }